norm + fixing things
