+++ includes/minishell.h
@@ -139,10 +139,12 @@ void	*sort_env(t_env	*env);
 /* cd.c */
 int		cd(t_input *in);
 
+/* cd_canon_form.c */
+char	*prep_path(char *var, char *path);
+
 /* cd_utils.c */
 int		check_directory(char *var, char *path);
 int		pwds(t_input *in, char *path);
-char	*make_path(char *tab, char *path, char **temp);
 char	*check_len(char	*path, t_env *env);
 
 /* unset.c */
